requirements.txt
joblib==1.0.1
numpy==1.20.3
scikit-learn==0.24.2
scikit-surprise==1.1.1
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
